refactor(source): Rename the source Matcher to LineMatcher
