+++ src/Command/Debug/DumpAstCommand.php
@@ -48,8 +48,8 @@ use Infection\FileSystem\FileSystem;
 use Infection\Logger\Console\ConsoleLogger;
 use Infection\PhpParser\Visitor\AddIdToTraversedNodesVisitor\AddIdToTraversedNodesVisitor;
 use Infection\PhpParser\Visitor\MarkTraversedNodesAsVisitedVisitor;
-use Infection\Source\Matcher\SimpleSourceLineMatcher;
-use Infection\Source\Matcher\SourceLineMatcher;
+use Infection\Source\Matcher\Line\SimpleSourceLineMatcher;
+use Infection\Source\Matcher\Line\SourceLineMatcher;
 use Infection\TestFramework\Tracing\Trace\EmptyTrace;
 use PhpParser\Node;
 use PhpParser\NodeTraverser;
