+++ src/Container/Container.php
@@ -151,9 +151,9 @@ use Infection\Source\Collector\MemoizedSourceCollector;
 use Infection\Source\Collector\SourceCollector;
 use Infection\Source\Collector\SourceCollectorFactory;
 use Infection\Source\Exception\NoSourceFound;
-use Infection\Source\Matcher\GitDiffSourceLineMatcher;
-use Infection\Source\Matcher\NullSourceLineMatcher;
-use Infection\Source\Matcher\SourceLineMatcher;
+use Infection\Source\Matcher\Line\GitDiffSourceLineMatcher;
+use Infection\Source\Matcher\Line\NullSourceLineMatcher;
+use Infection\Source\Matcher\Line\SourceLineMatcher;
 use Infection\Source\PreloadedSourceChecker;
 use Infection\StaticAnalysis\Config\StaticAnalysisConfigLocator;
 use Infection\StaticAnalysis\StaticAnalysisToolAdapter;
