+++ src/Source/Matcher/Line/SourceLineMatcher.php
@@ -33,17 +33,26 @@
 
 declare(strict_types=1);
 
-namespace Infection\Source\Matcher;
+namespace Infection\Source\Matcher\Line;
 
-use DomainException;
+use Infection\Source\Exception\NoSourceFound;
 
 /**
+ * Determines whether a specific line range in a file matches certain criteria.
+ *
  * @internal
  */
-final readonly class FakeSourceLineMatcher implements SourceLineMatcher
+interface SourceLineMatcher
 {
-    public function touches(string $fileRealPath, int $startLine, int $endLine): bool
-    {
-        throw new DomainException('Not implemented');
-    }
+    /**
+     * Checks whether the specified line range in the given file matches the
+     * criteria defined by this matcher.
+     *
+     * @param string $fileRealPath Absolute path to the file to check.
+     * @param positive-int $startLine Starting line number of the range (inclusive).
+     * @param positive-int $endLine Ending line number of the range (inclusive).
+     *
+     * @throws NoSourceFound
+     */
+    public function touches(string $fileRealPath, int $startLine, int $endLine): bool;
 }
