+++ tests/phpunit/Mutator/MutatorRobustnessTest.php
@@ -38,8 +38,8 @@ namespace Infection\Tests\Mutator;
 use Infection\Mutation\FileMutationGenerator;
 use Infection\Mutator\Mutator;
 use Infection\Mutator\ProfileList;
-use Infection\Source\Matcher\NullSourceLineMatcher;
-use Infection\Source\Matcher\SourceLineMatcher;
+use Infection\Source\Matcher\Line\NullSourceLineMatcher;
+use Infection\Source\Matcher\Line\SourceLineMatcher;
 use Infection\TestFramework\Tracing\Tracer;
 use Infection\Testing\MutatorName;
 use Infection\Testing\SingletonContainer;
