+++ tests/phpunit/PhpParser/Visitor/EnrichmentTraverse/EnrichmentTraverseIntegrationTest.php
@@ -39,9 +39,9 @@ use Infection\AbstractTestFramework\Coverage\TestLocation;
 use Infection\Container\Container;
 use Infection\Differ\ChangedLinesRange;
 use Infection\PhpParser\Visitor\MarkTraversedNodesAsVisitedVisitor;
-use Infection\Source\Matcher\NullSourceLineMatcher;
-use Infection\Source\Matcher\SimpleSourceLineMatcher;
-use Infection\Source\Matcher\SourceLineMatcher;
+use Infection\Source\Matcher\Line\NullSourceLineMatcher;
+use Infection\Source\Matcher\Line\SimpleSourceLineMatcher;
+use Infection\Source\Matcher\Line\SourceLineMatcher;
 use Infection\TestFramework\Tracing\Trace\Trace;
 use Infection\Testing\FileSystem\MockSplFileInfo;
 use Infection\Testing\SingletonContainer;
