+++ tests/phpunit/PhpParser/Visitor/ExcludeUnchangedLinesVisitor/ExcludeUnchangedLinesVisitorTest.php
@@ -42,8 +42,8 @@ use Infection\Differ\ChangedLinesRange;
 use Infection\PhpParser\Visitor\ExcludeUnchangedLinesVisitor;
 use Infection\PhpParser\Visitor\LabelNodesAsEligibleVisitor;
 use Infection\PhpParser\Visitor\MarkTraversedNodesAsVisitedVisitor;
-use Infection\Source\Matcher\FakeSourceLineMatcher;
-use Infection\Source\Matcher\SimpleSourceLineMatcher;
+use Infection\Source\Matcher\Line\FakeSourceLineMatcher;
+use Infection\Source\Matcher\Line\SimpleSourceLineMatcher;
 use Infection\Tests\PhpParser\Visitor\VisitorTestCase\VisitorTestCase;
 use PhpParser\Node;
 use PhpParser\Node\Name;
